chore: increase test coverage; reduce flaky, CI improvements
* Increasing test coverage of overall code
* Reduce flakiness on certain tests
* Simplified and removed some hardcoded variables
* Partition CI for tests
* Properly cache docker images
* Cleanup actions and user blacksmith runners
