+++ .github/workflows/lint.yml
@@ -46,12 +46,6 @@ jobs:
 
       - name: Install dependencies
         run: mix deps.get
-      - name: Set up Postgres
-        run: docker compose -f docker-compose.dbs.yml up -d
-      - name: Run main database migrations
-        run: mix ecto.migrate --log-migrator-sql
-      - name: Run database tenant migrations
-        run: mix ecto.migrate --migrations-path lib/realtime/tenants/repo/migrations
       - name: Check for warnings
         run: mix compile --force --warnings-as-errors
       - name: Run format check
@@ -77,5 +71,3 @@ jobs:
           mix dialyzer.build
       - name: Run dialyzer
         run: mix dialyzer
-      - name: Run dev seeds
-        run: DB_ENC_KEY="1234567890123456" mix ecto.setup
