+++ test/realtime/tenants/cache_test.exs
@@ -1,5 +1,4 @@
 defmodule Realtime.Tenants.CacheTest do
-  # async: false due to the usage of dev_realtime tenant
   use Realtime.DataCase, async: false
 
   alias Realtime.Api
@@ -61,71 +60,115 @@ defmodule Realtime.Tenants.CacheTest do
   describe "distributed_invalidate_tenant_cache/1" do
     setup do
       {:ok, node} = Clustered.start()
-      %{node: node}
+
+      tenant =
+        Ecto.Adapters.SQL.Sandbox.unboxed_run(Realtime.Repo, fn ->
+          tenant_fixture()
+        end)
+
+      on_exit(fn ->
+        Ecto.Adapters.SQL.Sandbox.unboxed_run(Realtime.Repo, fn ->
+          Realtime.Api.delete_tenant_by_external_id(tenant.external_id)
+        end)
+      end)
+
+      %{node: node, tenant: tenant}
     end
 
-    test "invalidates the cache given a tenant_id", %{node: node} do
-      external_id = "dev_tenant"
-      %Api.Tenant{name: expected_name} = tenant = Tenants.get_tenant_by_external_id(external_id)
-
+    test "invalidates the cache given a tenant_id", %{node: node, tenant: tenant} do
+      external_id = tenant.external_id
+      expected_name = tenant.name
       dummy_name = random_string()
+      dummy_tenant = %{tenant | name: dummy_name}
 
-      # Ensure cache has the values
-      Realtime.Tenants.Cache.update_cache(%{tenant | name: dummy_name})
-
-      Rpc.enhanced_call(node, Realtime.Tenants.Cache, :update_cache, [%{tenant | name: dummy_name}])
-
-      # Cache showing old value
-      assert {:ok, %Api.Tenant{name: ^dummy_name}} = Cachex.get(Cache, {:get_tenant_by_external_id, external_id})
+      assert {:ok, true} = Cache.update_cache(dummy_tenant)
 
       assert {:ok, %Api.Tenant{name: ^dummy_name}} =
-               Rpc.enhanced_call(node, Cachex, :get, [Cache, {:get_tenant_by_external_id, external_id}])
+               Cachex.get(Cache, {:get_tenant_by_external_id, external_id})
 
-      # Invalidate cache
+      seed_remote_cache(node, external_id, dummy_tenant)
+
       assert :ok = Cache.distributed_invalidate_tenant_cache(external_id)
 
-      # wait for cache to be invalidated in both nodes
-      Process.sleep(200)
-      # Cache showing new value
-      assert %Api.Tenant{name: ^expected_name} = Cache.get_tenant_by_external_id(external_id)
+      assert_eventually(fn ->
+        %Api.Tenant{name: ^expected_name} = Cache.get_tenant_by_external_id(external_id)
 
-      assert %Api.Tenant{name: ^expected_name} =
-               Rpc.enhanced_call(node, Cache, :get_tenant_by_external_id, [external_id])
+        %Api.Tenant{name: ^expected_name} =
+          Rpc.enhanced_call(node, Cache, :get_tenant_by_external_id, [external_id])
+      end)
     end
   end
 
   describe "global_cache_update/1" do
     setup do
       {:ok, node} = Clustered.start()
-      %{node: node}
+
+      tenant =
+        Ecto.Adapters.SQL.Sandbox.unboxed_run(Realtime.Repo, fn ->
+          tenant_fixture()
+        end)
+
+      on_exit(fn ->
+        Ecto.Adapters.SQL.Sandbox.unboxed_run(Realtime.Repo, fn ->
+          Realtime.Api.delete_tenant_by_external_id(tenant.external_id)
+        end)
+      end)
+
+      %{node: node, tenant: tenant}
     end
 
-    test "update the cache given a tenant_id", %{node: node} do
-      external_id = "dev_tenant"
-      %Api.Tenant{name: expected_name} = tenant = Tenants.get_tenant_by_external_id(external_id)
-
+    test "update the cache given a tenant_id", %{node: node, tenant: tenant} do
+      external_id = tenant.external_id
+      expected_name = tenant.name
       dummy_name = random_string()
+      dummy_tenant = %{tenant | name: dummy_name}
 
-      # Ensure cache has the values
-      Realtime.Tenants.Cache.update_cache(%{tenant | name: dummy_name})
+      assert {:ok, true} = Cache.update_cache(dummy_tenant)
 
-      Rpc.enhanced_call(node, Cache, :update_cache, [%{tenant | name: dummy_name}])
+      assert {:ok, %Api.Tenant{name: ^dummy_name}} =
+               Cachex.get(Cache, {:get_tenant_by_external_id, external_id})
 
-      # Cache showing old value
-      assert %Api.Tenant{name: ^dummy_name} = Cache.get_tenant_by_external_id(external_id)
-      assert %Api.Tenant{name: ^dummy_name} = Rpc.enhanced_call(node, Cache, :get_tenant_by_external_id, [external_id])
+      seed_remote_cache(node, external_id, dummy_tenant)
 
-      # Update cache
       assert :ok = Cache.global_cache_update(tenant)
 
-      # wait for cache to be updated in both nodes
-      Process.sleep(200)
+      assert_eventually(fn ->
+        {:ok, %Api.Tenant{name: ^expected_name}} =
+          Cachex.get(Cache, {:get_tenant_by_external_id, external_id})
 
-      # Cache showing new value
-      assert {:ok, %Api.Tenant{name: ^expected_name}} = Cachex.get(Cache, {:get_tenant_by_external_id, external_id})
-
-      assert {:ok, %Api.Tenant{name: ^expected_name}} =
-               Rpc.enhanced_call(node, Cachex, :get, [Cache, {:get_tenant_by_external_id, external_id}])
+        {:ok, %Api.Tenant{name: ^expected_name}} =
+          Rpc.enhanced_call(node, Cachex, :get, [Cache, {:get_tenant_by_external_id, external_id}])
+      end)
     end
+  end
+
+  defp seed_remote_cache(node, external_id, tenant, attempts \\ 20) do
+    Rpc.enhanced_call(node, Cache, :update_cache, [tenant])
+
+    case Rpc.enhanced_call(node, Cachex, :get, [Cache, {:get_tenant_by_external_id, external_id}]) do
+      {:ok, %Api.Tenant{external_id: ^external_id, name: name}} when name == tenant.name ->
+        :ok
+
+      _other when attempts > 0 ->
+        Process.sleep(50)
+        seed_remote_cache(node, external_id, tenant, attempts - 1)
+
+      other ->
+        flunk("Failed to seed remote cache after retries, last result: #{inspect(other)}")
+    end
+  end
+
+  defp assert_eventually(fun, attempts \\ 50, interval \\ 100)
+
+  defp assert_eventually(fun, 0, _interval) do
+    fun.()
+  end
+
+  defp assert_eventually(fun, attempts, interval) do
+    fun.()
+  rescue
+    _ ->
+      Process.sleep(interval)
+      assert_eventually(fun, attempts - 1, interval)
   end
 end
