+++ test/realtime_web/channels/realtime_channel_test.exs
@@ -14,6 +14,9 @@ defmodule RealtimeWeb.RealtimeChannelTest do
 
   setup do
     tenant = Containers.checkout_tenant(run_migrations: true)
+    {:ok, db_conn} = Realtime.Database.connect(tenant, "realtime_test", :stop)
+    Integrations.setup_postgres_changes(db_conn)
+    GenServer.stop(db_conn)
     Realtime.Tenants.Cache.update_cache(tenant)
     {:ok, tenant: tenant}
   end
